Victory Conditions and Disabling Regions (#15)
Update scenarios so they can honor their specified victory types,
rather than being forced into using what is selected in the lobby.

Revert the changes to 'World Domination - All' because we
can enforce that on the custom scenarios without breaking
the defaults.

Update all 4 scenarios to use the custom victory type.

Allow regions to be disabled if they should not be included
in the campaign.

Enforce that start regions are not accidentally disabled
or disallowed in a scenario.

Add the ability to set a TeamVictoryCondition.
A team can win by controlling certain territories for a given
number of turns. By default, controlling for one turn means
ending the turn owning the territories, so no extra turn
of holding the territories.